+++ src/app/(auth)/OAuthSignInComponent.tsx
@@ -1,0 +1,92 @@
+import { Separator } from "@/components/ui/separator";
+import { Button } from "@/components/ui/button";
+import React from "react";
+import { signIn } from "next-auth/react";
+
+function OAuthSignInComponent() {
+  const oauthSigin = async (provider: string) => {
+    await signIn(provider, {
+      callbackUrl: "/dashboard",
+    });
+  };
+  return (
+    <>
+      <div className="flex justify-center">
+        <Separator className="my-4 rounded h-0.5 w-5/12" />
+        <div className="pt-1.5 pl-2 pr-2 text-sm text-zinc-500">or</div>
+        <Separator className="my-4 rounded h-0.5 w-5/12" />
+      </div>
+      <Button
+        onClick={() => {
+          oauthSigin("google");
+        }}
+        className="w-full"
+      >
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          x="0px"
+          y="0px"
+          width="35"
+          height="25"
+          viewBox="0 0 48 48"
+        >
+          <path
+            fill="#fbc02d"
+            d="M43.611,20.083H42V20H24v8h11.303c-1.649,4.657-6.08,8-11.303,8c-6.627,0-12-5.373-12-12	s5.373-12,12-12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C12.955,4,4,12.955,4,24s8.955,20,20,20	s20-8.955,20-20C44,22.659,43.862,21.35,43.611,20.083z"
+          ></path>
+          <path
+            fill="#e53935"
+            d="M6.306,14.691l6.571,4.819C14.655,15.108,18.961,12,24,12c3.059,0,5.842,1.154,7.961,3.039	l5.657-5.657C34.046,6.053,29.268,4,24,4C16.318,4,9.656,8.337,6.306,14.691z"
+          ></path>
+          <path
+            fill="#4caf50"
+            d="M24,44c5.166,0,9.86-1.977,13.409-5.192l-6.19-5.238C29.211,35.091,26.715,36,24,36	c-5.202,0-9.619-3.317-11.283-7.946l-6.522,5.025C9.505,39.556,16.227,44,24,44z"
+          ></path>
+          <path
+            fill="#1565c0"
+            d="M43.611,20.083L43.595,20L42,20H24v8h11.303c-0.792,2.237-2.231,4.166-4.087,5.571	c0.001-0.001,0.002-0.001,0.003-0.002l6.19,5.238C36.971,39.205,44,34,44,24C44,22.659,43.862,21.35,43.611,20.083z"
+          ></path>
+        </svg>
+        Sign In with Google
+      </Button>
+      <Button
+        onClick={() => {
+          oauthSigin("github");
+        }}
+        className="w-full"
+      >
+        <svg
+          width="35"
+          height="25"
+          viewBox="0 0 20 20"
+          version="1.1"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <g
+            id="Page-1"
+            stroke="none"
+            stroke-width="1"
+            fill="none"
+            fill-rule="evenodd"
+          >
+            <g
+              id="Dribbble-Light-Preview"
+              transform="translate(-140.000000, -7559.000000)"
+              fill="#ffffff"
+            >
+              <g id="icons" transform="translate(56.000000, 160.000000)">
+                <path
+                  d="M94,7399 C99.523,7399 104,7403.59 104,7409.253 C104,7413.782 101.138,7417.624 97.167,7418.981 C96.66,7419.082 96.48,7418.762 96.48,7418.489 C96.48,7418.151 96.492,7417.047 96.492,7415.675 C96.492,7414.719 96.172,7414.095 95.813,7413.777 C98.04,7413.523 100.38,7412.656 100.38,7408.718 C100.38,7407.598 99.992,7406.684 99.35,7405.966 C99.454,7405.707 99.797,7404.664 99.252,7403.252 C99.252,7403.252 98.414,7402.977 96.505,7404.303 C95.706,7404.076 94.85,7403.962 94,7403.958 C93.15,7403.962 92.295,7404.076 91.497,7404.303 C89.586,7402.977 88.746,7403.252 88.746,7403.252 C88.203,7404.664 88.546,7405.707 88.649,7405.966 C88.01,7406.684 87.619,7407.598 87.619,7408.718 C87.619,7412.646 89.954,7413.526 92.175,7413.785 C91.889,7414.041 91.63,7414.493 91.54,7415.156 C90.97,7415.418 89.522,7415.871 88.63,7414.304 C88.63,7414.304 88.101,7413.319 87.097,7413.247 C87.097,7413.247 86.122,7413.234 87.029,7413.87 C87.029,7413.87 87.684,7414.185 88.139,7415.37 C88.139,7415.37 88.726,7417.2 91.508,7416.58 C91.513,7417.437 91.522,7418.245 91.522,7418.489 C91.522,7418.76 91.338,7419.077 90.839,7418.982 C86.865,7417.627 84,7413.783 84,7409.253 C84,7403.59 88.478,7399 94,7399"
+                  id="github-[#142]"
+                ></path>
+              </g>
+            </g>
+          </g>
+        </svg>
+        Sign In with GitHub
+      </Button>
+    </>
+  );
+}
+
+export default OAuthSignInComponent;
